+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.02](https://github.com/realityforge/revapi-diff/tree/v0.02) (2019-02-09)
 [Full Changelog](https://github.com/realityforge/revapi-diff/compare/v0.01...v0.02)