+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.03](https://github.com/realityforge/revapi-diff/tree/v0.03) (2019-02-11)
+[Full Changelog](https://github.com/realityforge/revapi-diff/compare/v0.02...v0.03)
 
 * Add an option `--expect-no-differences` that will generate a non-zero exit code when differences are
   detected otherwise the tool has been updated to return `0` on success.